--- a/administrator/components/com_users/helpers/users.php
+++ b/administrator/components/com_users/helpers/users.php
@@ -130,31 +130,12 @@ class UsersHelper
 	 */
 	public static function getGroups()
 	{
-		$db = JFactory::getDbo();
-		$query = $db->getQuery(true)
-			->select('a.id AS value')
-			->select('a.title AS text')
-			->select('COUNT(DISTINCT b.id) AS level')
-			->from('#__usergroups as a')
-			->join('LEFT', '#__usergroups  AS b ON a.lft > b.lft AND a.rgt < b.rgt')
-			->group('a.id, a.title, a.lft, a.rgt')
-			->order('a.lft ASC');
-		$db->setQuery($query);
-
-		try
-		{
-			$options = $db->loadObjectList();
-		}
-		catch (RuntimeException $e)
-		{
-			JError::raiseNotice(500, $e->getMessage());
-
-			return null;
-		}
+		$options = JHelperUsergroups::getInstance()->getAll();
 
 		foreach ($options as &$option)
 		{
-			$option->text = str_repeat('- ', $option->level) . $option->text;
+			$option->value = $option->id;
+			$option->text = str_repeat('- ', $option->level) . $option->title;
 		}
 
 		return $options;