--- a/administrator/components/com_users/models/fields/groupparent.php
+++ b/administrator/components/com_users/models/fields/groupparent.php
@@ -35,36 +35,17 @@ class JFormFieldGroupParent extends JFormFieldList
 	 */
 	protected function getOptions()
 	{
-		$options = array();
+		$options = JHelperUsergroups::getInstance()->getAll();
 
-		$db = JFactory::getDbo();
 		$user = JFactory::getUser();
-		$query = $db->getQuery(true)
-			->select('a.id AS value, a.title AS text, COUNT(DISTINCT b.id) AS level')
-			->from('#__usergroups AS a')
-			->join('LEFT', $db->quoteName('#__usergroups') . ' AS b ON a.lft > b.lft AND a.rgt < b.rgt');
 
 		// Prevent parenting to children of this item.
 		if ($id = $this->form->getValue('id'))
 		{
-			$query->join('LEFT', $db->quoteName('#__usergroups') . ' AS p ON p.id = ' . (int) $id)
-				->where('NOT(a.lft >= p.lft AND a.rgt <= p.rgt)');
+			unset($options[$id]);
 		}
 
-		$query->group('a.id, a.title, a.lft, a.rgt')
-			->order('a.lft ASC');
-
-		// Get the options.
-		$db->setQuery($query);
-
-		try
-		{
-			$options = $db->loadObjectList();
-		}
-		catch (RuntimeException $e)
-		{
-			JError::raiseWarning(500, $e->getMessage());
-		}
+		$options = array_values($options);
 
 		// Pad the option text with spaces using depth level as a multiplier.
 		for ($i = 0, $n = count($options); $i < $n; $i++)
@@ -72,7 +53,8 @@ class JFormFieldGroupParent extends JFormFieldList
 			// Show groups only if user is super admin or group is not super admin
 			if ($user->authorise('core.admin') || (!JAccess::checkGroup($options[$i]->value, 'core.admin')))
 			{
-				$options[$i]->text = str_repeat('- ', $options[$i]->level) . $options[$i]->text;
+				$options[$i]->value = $options[$i]->id;
+				$options[$i]->text = str_repeat('- ', $options[$i]->level) . $options[$i]->title;
 			}
 			else
 			{