--- a/administrator/components/com_users/models/groups.php
+++ b/administrator/components/com_users/models/groups.php
@@ -93,8 +93,6 @@ class UsersModelGroups extends JModelList
 	 */
 	public function getItems()
 	{
-		$db = $this->getDbo();
-
 		// Get a storage key.
 		$store = $this->getStoreId();
 
@@ -111,58 +109,15 @@ class UsersModelGroups extends JModelList
 				return $items;
 			}
 
-			// First pass: get list of the group id's and reset the counts.
-			$groupIds = array();
-
-			foreach ($items as $item)
-			{
-				$groupIds[] = (int) $item->id;
-			}
-
-			// Get total enabled users in group.
-			$query = $db->getQuery(true);
-
-			// Count the objects in the user group.
-			$query->select('map.group_id, COUNT(DISTINCT map.user_id) AS user_count')
-				->from($db->quoteName('#__user_usergroup_map', 'map'))
-				->join('LEFT', $db->quoteName('#__users', 'u') . ' ON ' . $db->quoteName('u.id') . ' = ' . $db->quoteName('map.user_id'))
-				->where($db->quoteName('map.group_id') . ' IN (' . implode(',', $groupIds) . ')')
-				->where($db->quoteName('u.block') . ' = 0')
-				->group($db->quoteName('map.group_id'));
-			$db->setQuery($query);
-
 			try
 			{
-				$countEnabled = $db->loadAssocList('group_id', 'count_enabled');
+				$items = $this->populateExtraData($items);
 			}
 			catch (RuntimeException $e)
 			{
 				$this->setError($e->getMessage());
+
 				return false;
-			}
-
-			// Get total disabled users in group.
-			$query->clear('where')
-				->where('map.group_id IN (' . implode(',', $groupIds) . ')')
-				->where('u.block = 1');
-			$db->setQuery($query);
-
-			try
-			{
-				$countDisabled = $db->loadAssocList('group_id', 'count_disabled');
-			}
-			catch (RuntimeException $e)
-			{
-				$this->setError($e->getMessage());
-				return false;
-			}
-
-			// Inject the values back into the array.
-			foreach ($items as $item)
-			{
-				$item->count_enabled   = isset($countEnabled[$item->id]) ? (int) $countEnabled[$item->id]['user_count'] : 0;
-				$item->count_disabled  = isset($countDisabled[$item->id]) ? (int) $countDisabled[$item->id]['user_count'] : 0;
-				$item->user_count      = $item->count_enabled + $item->count_disabled;
 			}
 
 			// Add the items to the internal cache.
@@ -192,11 +147,6 @@ class UsersModelGroups extends JModelList
 		);
 		$query->from($db->quoteName('#__usergroups') . ' AS a');
 
-		// Add the level in the tree.
-		$query->select('COUNT(DISTINCT c2.id) AS level')
-			->join('LEFT OUTER', $db->quoteName('#__usergroups') . ' AS c2 ON a.lft > c2.lft AND a.rgt < c2.rgt')
-			->group('a.id, a.lft, a.rgt, a.parent_id, a.title');
-
 		// Filter the comments over the search string if set.
 		$search = $this->getState('filter.search');
 
@@ -217,5 +167,81 @@ class UsersModelGroups extends JModelList
 		$query->order($db->escape($this->getState('list.ordering', 'a.lft')) . ' ' . $db->escape($this->getState('list.direction', 'ASC')));
 
 		return $query;
+	}
+
+	/**
+	 * Populate level & path for items.
+	 *
+	 * @param   array  $items  Array of stdClass objects
+	 *
+	 * @return  array
+	 *
+	 * @since   DEPLOY_VERSION
+	 */
+	private function populateExtraData(array $items)
+	{
+		// First pass: get list of the group id's and reset the counts.
+		$groupsByKey = array();
+
+		foreach ($items as $item)
+		{
+			$groupsByKey[(int) $item->id] = $item;
+		}
+
+		$groupIds = array_keys($groupsByKey);
+
+		$db = $this->getDbo();
+
+		// Get total enabled users in group.
+		$query = $db->getQuery(true);
+
+		// Count the objects in the user group.
+		$query->select('map.group_id, COUNT(DISTINCT map.user_id) AS user_count')
+			->from($db->quoteName('#__user_usergroup_map', 'map'))
+			->join('LEFT', $db->quoteName('#__users', 'u') . ' ON ' . $db->quoteName('u.id') . ' = ' . $db->quoteName('map.user_id'))
+			->where($db->quoteName('map.group_id') . ' IN (' . implode(',', $groupIds) . ')')
+			->where($db->quoteName('u.block') . ' = 0')
+			->group($db->quoteName('map.group_id'));
+		$db->setQuery($query);
+
+		try
+		{
+			$countEnabled = $db->loadAssocList('group_id', 'count_enabled');
+		}
+		catch (RuntimeException $e)
+		{
+			$this->setError($e->getMessage());
+
+			return false;
+		}
+
+		// Get total disabled users in group.
+		$query->clear('where')
+			->where('map.group_id IN (' . implode(',', $groupIds) . ')')
+			->where('u.block = 1');
+		$db->setQuery($query);
+
+		try
+		{
+			$countDisabled = $db->loadAssocList('group_id', 'count_disabled');
+		}
+		catch (RuntimeException $e)
+		{
+			$this->setError($e->getMessage());
+
+			return false;
+		}
+
+		// Inject the values back into the array.
+		foreach ($groupsByKey as &$item)
+		{
+			$item->count_enabled   = isset($countEnabled[$item->id]) ? (int) $countEnabled[$item->id]['user_count'] : 0;
+			$item->count_disabled  = isset($countDisabled[$item->id]) ? (int) $countDisabled[$item->id]['user_count'] : 0;
+			$item->user_count      = $item->count_enabled + $item->count_disabled;
+		}
+
+		$groups = new JHelperUsergroups($groupsByKey);
+
+		return array_values($groups->getAll());
 	}
 }