--- a/libraries/cms/html/access.php
+++ b/libraries/cms/html/access.php
@@ -92,19 +92,12 @@ abstract class JHtmlAccess
 	 */
 	public static function usergroup($name, $selected, $attribs = '', $allowAll = true, $id = false)
 	{
-		$db = JFactory::getDbo();
-		$query = $db->getQuery(true)
-			->select('a.id AS value, a.title AS text, COUNT(DISTINCT b.id) AS level')
-			->from($db->quoteName('#__usergroups') . ' AS a')
-			->join('LEFT', $db->quoteName('#__usergroups') . ' AS b ON a.lft > b.lft AND a.rgt < b.rgt')
-			->group('a.id, a.title, a.lft, a.rgt')
-			->order('a.lft ASC');
-		$db->setQuery($query);
-		$options = $db->loadObjectList();
+		$options = array_values(JHelperUsergroups::getInstance()->getAll());
 
 		for ($i = 0, $n = count($options); $i < $n; $i++)
 		{
-			$options[$i]->text = str_repeat('- ', $options[$i]->level) . $options[$i]->text;
+			$options[$i]->value = $options[$i]->id;
+			$options[$i]->text = str_repeat('- ', $options[$i]->level) . $options[$i]->title;
 		}
 
 		// If all usergroups is allowed, push it into the array.