--- a/libraries/cms/html/user.php
+++ b/libraries/cms/html/user.php
@@ -27,19 +27,12 @@ abstract class JHtmlUser
 	 */
 	public static function groups($includeSuperAdmin = false)
 	{
-		$db = JFactory::getDbo();
-		$query = $db->getQuery(true)
-			->select('a.id AS value, a.title AS text, COUNT(DISTINCT b.id) AS level')
-			->from($db->quoteName('#__usergroups') . ' AS a')
-			->join('LEFT', $db->quoteName('#__usergroups') . ' AS b ON a.lft > b.lft AND a.rgt < b.rgt')
-			->group('a.id, a.title, a.lft, a.rgt')
-			->order('a.lft ASC');
-		$db->setQuery($query);
-		$options = $db->loadObjectList();
+		$options = array_values(JHelperUsergroups::getInstance()->getAll());
 
 		for ($i = 0, $n = count($options); $i < $n; $i++)
 		{
-			$options[$i]->text = str_repeat('- ', $options[$i]->level) . $options[$i]->text;
+			$options[$i]->value = $options[$i]->id;
+			$options[$i]->text = str_repeat('- ', $options[$i]->level) . $options[$i]->title;
 			$groups[] = JHtml::_('select.option', $options[$i]->value, $options[$i]->text);
 		}