--- a/libraries/joomla/form/fields/rules.php
+++ b/libraries/joomla/form/fields/rules.php
@@ -406,6 +406,7 @@ class JFormFieldRules extends JFormField
 						$result['class'] = 'label label-important';
 						$result['text']  = JText::_('JLIB_RULES_NOT_ALLOWED_DEFAULT');
 					}
+
 					/**
 					 * Component/Item with explicit "Denied" permission at parent Asset (Category, Component or Global config) configuration.
 					 * Or some parent group has an explicit "Denied".
@@ -454,16 +455,14 @@ class JFormFieldRules extends JFormField
 	 */
 	protected function getUserGroups()
 	{
-		$db = JFactory::getDbo();
-		$query = $db->getQuery(true)
-			->select('a.id AS value, a.title AS text, COUNT(DISTINCT b.id) AS level, a.parent_id')
-			->from('#__usergroups AS a')
-			->join('LEFT', $db->quoteName('#__usergroups') . ' AS b ON a.lft > b.lft AND a.rgt < b.rgt')
-			->group('a.id, a.title, a.lft, a.rgt, a.parent_id')
-			->order('a.lft ASC');
-		$db->setQuery($query);
-		$options = $db->loadObjectList();
+		$options = JHelperUsergroups::getInstance()->getAll();
 
-		return $options;
+		foreach ($options as &$option)
+		{
+			$option->value = $option->id;
+			$option->text  = $option->title;
+		}
+
+		return array_values($options);
 	}
 }