@zero-24 2016-08-31
Result.

- [x] Go to backend users view and ensure that batch works.
- [x] Go to backend users view and check that group searchtools filter works properly.
- [x] Go to backend users view and create an user. Confirm that groups selection is correctly saved.
- [x] Go to backend groups view and ensure that groups are shown.
- [x] Go to backend groups view and ensure that pagination works. (tip set the dropdown to 5)
- [x] Go to backend groups view and ensure that groups are shown. (that looks dublicated? but success :smile: )
- [x] Go to backend groups view and try to create a new group
- [x] Go to backend groups view and edit a group. Ensure that the group is not shown in the parent field.
- [x] Go to backend groups view and edit a group to change the parent group. Confirm data is saved.
- [x] Go to backend articles view and edit an article to change its permissions. Confirm data is saved.

note: I have not exiplite tested the performance.