@dgt41 2016-08-31
I have tested this item :white_check_mark: successfully on 3f1c170ca3c48406361f658de60bcf156e2c6647<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/11853">issues.joomla.org/joomla-cms/11853</a>.</sub>