@zero-24 2016-08-31
I have broken travis :( If that is fixed we can set it RTC. https://github.com/phproberto/joomla-cms/pull/5 Thanks :+1: