@zero-24 2016-08-31
I have tested this item :white_check_mark: successfully on 2d152093a8ac735be0b6c2e8bbc79060a243771d<br /><br />Thanks. Travis is happy now. RTC<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/11853">issues.joomla.org/joomla-cms/11853</a>.</sub>